pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-deep-dive"
version = "0.1.0"
description = "Local coding agent prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2,<3",
  "rich>=13,<15",
  "typer>=0.12,<1",
  "litellm>=1.63,<2",
  "python-frontmatter>=1.1.0",
  "croniter>=1.4",
  "mcp>=1.0,<2",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
]

[project.scripts]
agent = "agent.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
